The @Transactional annotation ensures that all operations are performed atomically. In other words, if the creation of the document fails, the creation of the attachment will also be reversed.
By marking the create method with the @Transactional annotation, both the attachment and the document will be saved within the same transaction, and if there is a problem when saving the document, the rollback will be done automatically, also reverting the creation of the attachment.

Before this annotation, even if an exception occurred when creating a new document, the attachment was being created inadvertently.